CT Skull Stripping within R

Skull Stripping (using FSL's BET) a CT file using fslr functions

CT_Skull_Strip(
  img,
  outfile = NULL,
  keepmask = TRUE,
  maskfile = NULL,
  inskull_mesh = FALSE,
  retimg = TRUE,
  reorient = FALSE,
  intern = TRUE,
  betcmd = "bet2",
  opts = "-f 0.01 -v",
  presmooth = TRUE,
  remask = TRUE,
  refill = FALSE,
  refill.thresh = 0.75,
  sigma = 1,
  lthresh = 0,
  uthresh = 100,
  smooth_before_threshold = FALSE,
  verbose = TRUE,
  ...
)

Arguments

img
(character) File to be skull stripped or object of class nifti

outfile
(character) output filename

keepmask
(logical) Should we keep the mask?

maskfile
(character) Filename for mask (if keepmask = TRUE). If NULL, then will do paste0(outfile, "_Mask").

inskull_mesh
(logical) Create inskull_mesh file from bet? (Warning - will take longer) This an exterior surface of the brain. (experimental) Also, if outfile is NULL, then this will be created in a temporary directory and not be retrieved.

retimg
(logical) return image of class nifti

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readNIfTI.

intern
(logical) pass to system

betcmd
(character) bet command to be used, see fslbet

opts
(character) additional options to fslbet

presmooth
(logical) indicator if pre-smoothing should be done before BET

remask
(logical) Mask the smoothed image with HU mask from initial step?

refill
(logical) indicator to post-smooth mask and then fill

refill.thresh
(numeric) Value to threshold post-smoothed mask

sigma
(integer) size of Gaussian kernel passed to fslsmooth if presmooth is TRUE

lthresh
(default: 0) Lower value to threshold CT fslthresh

uthresh
(default: 100) Upper value to threshold CT fslthresh

smooth_before_threshold
Should the image be smoothed before thresholding? This can be useful for bone-window scans.

verbose
(logical) Should diagnostic output be printed?

...
additional arguments passed to fslbet.

Value

character or logical depending on intern