Preprocess data for ICH Segmentation

Will do skull stripping and registration to the CT template

ich_preprocess(
  img,
  skull_strip = TRUE,
  robust = TRUE,
  mask = NULL,
  n4_correct = FALSE,
  template.file = system.file("scct_unsmooth_SS_0.01.nii.gz", package = "ichseg"),
  outprefix = NULL,
  typeofTransform = c("Rigid", "Affine"),
  interpolator = "Linear",
  outfile = NULL,
  verbose = TRUE,
  shiny = FALSE,
  roi = NULL,
  ...
)

ich_cnn_preprocess(
  ...,
  template.file = system.file("scct_unsmooth_SS_0.01_128x128x128.nii.gz", package =
    "ichseg")
)

Arguments

img
CT image, object of class nifti or character filename

skull_strip
Should the image be skull stripped? If not, a mask must be specified. If mask specified, will override this option

robust
If skull stripping, should CT_Skull_Strip_robust be used vs. CT_Skull_Strip

mask
binary brain mask, object of class nifti or character filename

n4_correct
Should N4 bias-field correction be done after skull-stripping

template.file
Template to register to (Skull stripped template)

outprefix
Passed to registration for the output transformations

typeofTransform
Transformation to the template

interpolator
Interpolation to be done after transformation

outfile
Output filename for transformed registered image

verbose
Print diagnostic output

shiny
Should shiny progress be called?

roi
Filename of ROI, which will be transformed

...
Additional options passsed to CT_Skull_Strip_robust or CT_Skull_Strip

Value

List of output images and transformations