Make CT Predictors

Create a set of predictors for ICH segmentation for CT

make_predictors(
  img,
  mask,
  roi = NULL,
  nvoxels = 1,
  moments = 1:4,
  center = c(FALSE, TRUE, TRUE, TRUE),
  lthresh = 40,
  uthresh = 80,
  sigmas = c(5, 10, 20),
  save_imgs = TRUE,
  outdir = NULL,
  stub = NULL,
  overwrite = FALSE,
  template.file = system.file("scct_unsmooth_SS_0.01.nii.gz", package = "ichseg"),
  mean.img = system.file("Mean_Image.nii.gz", package = "ichseg"),
  sd.img = system.file("SD_Image.nii.gz", package = "ichseg"),
  zscore.typeofTransform = "SyN",
  zscore.interpolator = "Linear",
  flip.typeofTransform = "Affine",
  flip.interpolator = "LanczosWindowedSinc",
  low_thresh = 1e-13,
  verbose = TRUE,
  shiny = FALSE,
  erode_mask = TRUE,
  ...
)

Arguments

img
Filename of image intensities

mask
Filename of brain mask

roi
Filename of ROI for Y

nvoxels
Voxel neighborhood

moments
Moments of neighborhood to take

center
Center the moments

lthresh
Lower threshold for neighborhood setting

uthresh
Upper threshold for neighborhood setting

sigmas
Sigma values for Gaussian smoothed images (in mm)

save_imgs
Logical to save all images that are created as predictors

outdir
Output directory of saved images, needs to be set if save_imgs = TRUE

stub
Basename to write image names if save_imgs = TRUE

overwrite
If save_imgs is TRUE, then should the files be overwritten? If not, then files will be read instead instead of code being re-run.

template.file
Template to register to (CT Template)

mean.img
Mean image in template space for z-scoring

sd.img
SD image in template space for z-scoring

zscore.typeofTransform
type of transform for z-scoring

zscore.interpolator
type of interpolator for z-scoring

flip.typeofTransform
type of transform for flipped difference

flip.interpolator
type of interpolator for flipped difference

low_thresh
Threshold for forcing values to zero

verbose
Logical indicator if output messages should be printed

shiny
Should shiny progress be called?

erode_mask
Should the brain mask be eroded?

...
options passed to get_neighbors

Value

List of a data.frame of Predictors and set of indices to keep in mask and an empty nifti object for plotting. Also the number of voxels of the roi that were not in the mask